moved AltMenu to unused files and moved TextBox into separate class
